Split out service/client.rs
